libkmod: Support raw module signature information
Enable kmod_module_get_info() to obtain the raw data, such as the
signature and key_id in the signed module, instead of the formatted
string, and provide the kmod_module_info_get_value_n() function to
obtain this piece of data. String formatters should be implemented
by the caller tools.

Signed-off-by: Tianjia Zhang <tianjia.zhang@linux.alibaba.com>